requirements.txt
flask
flask-cors
python-doctr
torch
torchvision
opencv-python
numpy
